Reworked quic-specific code

+++ TLS-Core/src/main/java/de/rub/nds/tlsattacker/core/quic/packet/OneRTTPacket.java
@@ -39,7 +39,7 @@ public class OneRTTPacket extends QuicPacket {
 
     public OneRTTPacket(byte flags) {
         super(QuicPacketType.ONE_RTT_PACKET);
-        this.setProtectedFlags((byte) flags);
+        this.setProtectedFlags(flags);
         protectedHeaderHelper.write(flags);
         this.packetSecret = QuicCryptoSecrets.APPLICATION_SECRET;
     }
